[feat] 지도 호출 방식 paging에서 list로 변경
