+++ app/src/main/java/com/dnd_8th_4_android/wery/data/api/MapService.kt
@@ -1,5 +1,6 @@
 package com.dnd_8th_4_android.wery.data.api
 
+import ResponseMapFeedList
 import com.dnd_8th_4_android.wery.data.remote.model.map.*
 import retrofit2.http.*
 
@@ -16,8 +17,7 @@ interface MapService {
         @Query("startLongitude") startLongitude: Double,
         @Query("endLatitude") endLatitude: Double,
         @Query("endLongitude") endLongitude: Double,
-        @Query("page") page: Int? = 1,
-    ): ResponseMapFeedLis
+    ): ResponseMapFeedList
 
     @GET("/content/map/detail")
     suspend fun getFeedData(
