+++ app/src/main/java/com/dnd_8th_4_android/wery/data/remote/datasource/MapDataSourceImpl.kt
@@ -1,5 +1,6 @@
 package com.dnd_8th_4_android.wery.data.remote.datasource
 
+import ResponseMapFeedList
 import com.dnd_8th_4_android.wery.data.api.MapService
 import com.dnd_8th_4_android.wery.data.remote.model.map.*
 import javax.inject.Inject
@@ -20,9 +21,8 @@ class MapDataSourceImpl @Inject constructor(private val mapService: MapService) 
         startLongitude: Double,
         endLatitude: Double,
         endLongitude: Double,
-        page: Int
-    ): ResponseMapFeedLis {
-        return mapService.getFeedList(startLatitude, startLatitude, endLatitude, endLongitude, page)
+    ): ResponseMapFeedList {
+        return mapService.getFeedList(startLatitude, startLatitude, endLatitude, endLongitude)
     }
 
     override suspend fun getMapFeedDetail(location: String): ResponseMapFeedData {
