+++ app/src/main/java/com/dnd_8th_4_android/wery/data/remote/model/map/ResponseMapFeedLis.kt
@@ -1,23 +1,16 @@
-package com.dnd_8th_4_android.wery.data.remote.model.map
-
 import com.dnd_8th_4_android.wery.data.remote.model.BaseResponse
 
-data class ResponseMapFeedLis(
-    val data: Data,
+data class ResponseMapFeedList(
+    val data: List<ResultMapFeedData>,
 ) : BaseResponse() {
-    data class Data(
-        val content: List<Content>,
-        val totalPages: Int
-    ) {
-        data class Content(
-            val contentImageUrl: String,
-            val counts: Int,
-            val groupId: Int,
-            val id: Int,
-            val latitude: Double,
-            val location: String,
-            val longitude: Double,
-            val userId: Int
-        )
-    }
+    data class ResultMapFeedData(
+        val id: Int,
+        val groupId: Int,
+        val counts:Int,
+        val latitude: Double,
+        val longitude: Double,
+        val userId: Int,
+        val contentImageUrl:String,
+        val location:String
+    )
 }
