+++ app/src/main/java/com/dnd_8th_4_android/wery/data/repository/MapRepositoryImpl.kt
@@ -1,7 +1,11 @@
 package com.dnd_8th_4_android.wery.data.repository
 
+import ResponseMapFeedList
 import com.dnd_8th_4_android.wery.data.remote.datasource.MapDataSource
-import com.dnd_8th_4_android.wery.data.remote.model.map.*
+import com.dnd_8th_4_android.wery.data.remote.model.map.RequestMapMissionList
+import com.dnd_8th_4_android.wery.data.remote.model.map.ResponseMapFeedData
+import com.dnd_8th_4_android.wery.data.remote.model.map.ResponseMapMissionData
+import com.dnd_8th_4_android.wery.data.remote.model.map.ResponseMapMissionList
 import com.dnd_8th_4_android.wery.domain.repository.MapRepository
 import javax.inject.Inject
 
@@ -20,14 +24,12 @@ class MapRepositoryImpl @Inject constructor(private val mapDataSource: MapDataSo
         startLongitude: Double,
         endLatitude: Double,
         endLongitude: Double,
-        page: Int
-    ): ResponseMapFeedLis {
+    ): ResponseMapFeedList {
         return mapDataSource.getFeedList(
             startLatitude,
             startLatitude,
             endLatitude,
             endLongitude,
-            page
         )
     }
